requirements.txt
Django==4.1.7
ipython==8.12.0
django-extensions==3.2.1
psycopg2-binary==2.9.6
django-debug-toolbar==4.0.0
django-crispy-forms==2.0
crispy-bootstrap5==0.7
Pillow==9.5.0
django-filter==23.2
djangorestframework==3.14.0
Markdown==3.4.3
celery==5.2.7




